Combine curl object tests
Closes #420
